QMetaEnum: fix narrowing warnings related to valueToKey*()
valueToKey*() takes an int, and QMetaEnum internally assumes the
underlying type of an enum is not bigger than int.

Using static_cast as it's easier to search for it in the codebase if
needed.

Found by compiling with -Wshorten-64-to-32.

Drive-by change: remove qdoc \fn command for qt_QMetaEnum_debugOperator,
it's an internal method and defined right under the doc block (the
signature still said "int" but it's qint64).

+++ src/corelib/kernel/qmetatype.cpp
@@ -1951,13 +1951,13 @@ static bool convertFromEnum(QMetaType fromType, const void *from, QMetaType toTy
     QMetaEnum en = metaEnumFromType(fromType);
     if (en.isValid()) {
         if (en.isFlag()) {
-            const QByteArray keys = en.valueToKeys(ll);
+            const QByteArray keys = en.valueToKeys(static_cast<int>(ll));
             if (toType.id() == QMetaType::QString)
                 *static_cast<QString *>(to) = QString::fromUtf8(keys);
             else
                 *static_cast<QByteArray *>(to) = keys;
         } else {
-            const char *key = en.valueToKey(ll);
+            const char *key = en.valueToKey(static_cast<int>(ll));
             if (toType.id() == QMetaType::QString)
                 *static_cast<QString *>(to) = QString::fromUtf8(key);
             else
